+++ packages/middlewares/src/constants.ts
@@ -4,4 +4,3 @@ export const X_AUTH_STATUS = 'x-auth-status'
 
 export const TP_TK = 'TP_TK'
 export const TP_SE = 'TP_SE'
-export const X_SOTO_SESSION = 'x-soto-session'
