+++ packages/middlewares/src/index.ts
@@ -1,4 +1,10 @@
+import { chain, MiddlewareFactory } from './chain'
+import { refreshSessionMiddleware } from './refresh-session'
+
 export { chain } from './chain'
 export { oldIosCookiesMiddleware } from './old-ios-cookie'
-export { sessionCookieMiddleware } from './session-cookie'
+export { refreshSessionMiddleware } from './refresh-session'
 export { NEED_LOGIN_IDENTIFIER, X_AUTH_STATUS } from './constants'
+
+export const middlewareWithSessionRefresh = (functions: MiddlewareFactory[]) =>
+  chain([refreshSessionMiddleware, ...functions])
